# BE-109: Improve render performance