--- a/.circleci/config.yml
+++ b/.circleci/config.yml
@@ -1,8 +1,8 @@
 version: 2.1
-orbs:
-  react-native: react-native-community/react-native@1.2.1
 jobs:
   test:
+    docker:
+      - react-native-community/react-native@1.2.1
     steps:
       - run: yarn install
       - run: yarn test:app