--- a/.circleci/config.yml
+++ b/.circleci/config.yml
@@ -1,10 +1,11 @@
 version: 2.1
 jobs:
   build:
+    working_directory: ~/BibleEngine
     docker:
       - image: circleci/node:latest
     steps:
       - checkout
       - run: yarn install
-      - run: cd app && yarn install
-      - run: yarn test
+      - run: cd ~/BibleEngine/app && yarn install
+      - run: cd ~/BibleEngine/app && yarn test