--- a/.circleci/config.yml
+++ b/.circleci/config.yml
@@ -6,6 +6,15 @@ jobs:
       - image: circleci/node:latest
     steps:
       - checkout
-      - run: yarn install
+      - restore_cache:
+          name: Restore Yarn Package Cache
+          keys:
+            - yarn-packages-{{ checksum "yarn.lock" }}
+      - run: yarn install --frozen-lockfile
+      - save_cache:
+          name: Save Yarn Package Cache
+          key: yarn-packages-{{ checksum "yarn.lock" }}
+          paths:
+            - ~/.cache/yarn
       - run: cd ~/BibleEngine/app && yarn install
       - run: cd ~/BibleEngine/app && yarn test