--- a/app/src/StrongsWord.tsx
+++ b/app/src/StrongsWord.tsx
@@ -53,19 +53,10 @@ export default class StrongsWord extends React.PureComponent<Props, State> {
 
   async componentDidMount() {
     this.mounted = true;
-    setTimeout(() => {
-      this.setDictionaryEntries(this.props.strongs);
-    }, 1000);
   }
 
   componentWillUnmount() {
     this.mounted = false;
-  }
-
-  async componentWillReceiveProps(nextProps: Props) {
-    setTimeout(() => {
-      this.setDictionaryEntries(nextProps.strongs);
-    }, 1000);
   }
 
   async setDictionaryEntries(strongs: string[]) {
@@ -89,6 +80,7 @@ export default class StrongsWord extends React.PureComponent<Props, State> {
 
   onPress = () => {
     this.setState({ ...this.state, popoverIsVisible: true });
+    this.setDictionaryEntries(this.props.strongs);
     setTimeout(() => {
       this.setState({
         ...this.state,