--- a/importers/package.json
+++ b/importers/package.json
@@ -29,6 +29,7 @@
         "@types/rimraf": "^2.0.2",
         "copyfiles": "^2.1.0",
         "jest": "24.5.0",
+        "sqlite3": "^4.1.0",
         "ts-jest": "24.0.0",
         "ts-node": "8.0.3",
         "tslint": "5.14.0",