--- a/importers/scripts/buildSqliteDatabase.ts
+++ b/importers/scripts/buildSqliteDatabase.ts
@@ -3,7 +3,6 @@ import { StepLexiconImporter } from './../src/step-lexicon/importer';
 import { BeDatabaseCreator } from './../src/BeDatabaseCreator.class';
 
 const run = async () => {
-    console.log(process.env.DATABASE_URL)
   const creator = new BeDatabaseCreator({
     type: 'sqlite',
     database: 'bibles.db',
@@ -12,16 +11,6 @@ const run = async () => {
 
   creator.addImporter(SwordImporter, {
     sourcePath: 'src/osis-sword-module/data/ESV2016_th.zip'
-  });
-  creator.addImporter(SwordImporter, {
-    sourcePath: 'src/osis-sword-module/data/ChiUns.zip',
-    versionMeta: {
-      title: 'Chinese Union Version',
-      uid: 'CUVs',
-      copyrightShort: 'Asia Bible Society 2013',
-      language: 'zh',
-      hasStrongs: true
-    }
   });
   creator.addImporter(StepLexiconImporter)