--- a/importers/scripts/buildSqliteDatabase.ts
+++ b/importers/scripts/buildSqliteDatabase.ts
@@ -1,3 +1,4 @@
+import { V11nImporter } from './../src/v11n-rules/index';
 import { SwordImporter } from './../src/osis-sword-module/src/importer';
 import { StepLexiconImporter } from './../src/step-lexicon/importer';
 import { BeDatabaseCreator } from './../src/BeDatabaseCreator.class';
@@ -8,11 +9,11 @@ const run = async () => {
     database: 'bibles.db',
     dropSchema: true
   });
-
+  creator.addImporter(V11nImporter);
   creator.addImporter(SwordImporter, {
     sourcePath: 'src/osis-sword-module/data/ESV2016_th.zip'
   });
-  creator.addImporter(StepLexiconImporter)
+  creator.addImporter(StepLexiconImporter);
 
   await creator.createDatabase();
 };