--- a/package.json
+++ b/package.json
@@ -6,7 +6,8 @@
     "scripts": {
         "postinstall": "yarn run buildAll",
         "heroku": "node server/lib/heroku.js",
-        "buildAll": "yarn workspace @bible-engine/core run build && yarn workspace @bible-engine/importers run build && yarn workspace @bible-engine/server run generateApiClient && yarn workspace @bible-engine/server run build && yarn workspace @bible-engine/client run build"
+        "buildAll": "yarn workspace @bible-engine/core run build && yarn workspace @bible-engine/importers run build && yarn workspace @bible-engine/server run generateApiClient && yarn workspace @bible-engine/server run build && yarn workspace @bible-engine/client run build",
+        "test:app": "yarn workspace @bible-engine/core run test"
     },
     "workspaces": {
         "packages": [