## Summary

Right now, the app feels a little sluggish, so I profiled every feature using `console.time`, and took the average value over 5 iterations:

For Genesis 1:
Baseline render performance: 1168 ms
Strongs:  396 ms
Internet check: 200 ms
Custom fonts: 161 ms
Version check: 100 ms
Loading screen animation: 80 ms
Cross references: 37 ms
Sidebar: no cost
Search: no cost

I also threw in a CI integration.

## Issue Link

https://github.com/tyndale/BibleEngine/issues/109